+++ tests/files_test/Queue/TestQueue.php
@@ -3,13 +3,13 @@
 /**
  * @generated class generated using Vinci Console
  */
-class UserQueue extends Solital\Core\Queue\Queue
+class TestQueue extends Solital\Core\Queue\Queue
 {
 	/**
 	 * dispatch
 	 */
 	public function dispatch()
 	{
-		echo 'queue 2' . PHP_EOL;
+		echo 'queue 1' . PHP_EOL;
 	}
 }