+++ tests/files_test/cache/list1.cache.php
@@ -1,1 +1,0 @@
-{"nome":"Harvey Specter","email":"specter@pearsonhardman.com","expire_at":1648211802}