+++ tests/files_test/cache/list2.cache.php
@@ -1,1 +1,0 @@
-{"nome":"Louis Litt","email":"liitup@pearsonhardman.com","expire_at":1648211802}